// ai-crm-leads/src/pattern-learner/raindrop.gen.ts
// Code generated by raindrop build, DO NOT EDIT.
import { ActorNamespace, Ai, Annotation, App, Logger, MRNObject, ServiceStub, SmartBucket, SmartMemory, SmartSql, Tracer, VectorIndex,  } from '@liquidmetal-ai/raindrop-framework';

export interface Env {
  _raindrop: {
    app: App;
  };
  AI: Ai;
  AI_INSIGHTS: SmartMemory;
  annotation: Annotation<Omit<MRNObject, 'type' | 'applicationName' | 'versionId'>>;
  CONVERSION_PATTERNS: VectorIndex;
  CRM_API: ServiceStub<import('../crm-api/index.js').default>;
  CRM_DATABASE: SmartSql;
  LEAD_SCORER: ActorNamespace<import('../lead-scorer/index.js').LeadScorer>;
  logger: Logger;
  SALES_DOCUMENTS: SmartBucket;
  tracer: Tracer;
  USER_SESSIONS: ActorNamespace<import('../user-sessions/index.js').UserSessions>;
}